Discover config options through /models
Let's discover all the config options that are available in the /models
API call automatically. This commit removes all config options that do
not need to be specified by the end user.

The code related to models is moved to a separate package to ensure
clear separation between the code related to models and the rest of the
app.